backend/requirements.txt
# CellCount Backend Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# ML/Detection
ultralytics>=8.1.0
huggingface_hub>=0.20.0

# File handling
python-multipart>=0.0.6
pillow>=10.2.0

# Configuration
pydantic-settings>=2.1.0

# Development (optional)
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# httpx>=0.26.0
